Separate structured light correspondence acquisition example and calibration example
